requirements.txt
fastapi==0.116.1
openai==1.100.2
pydantic==2.11.7
python-dotenv==1.1.1
supabase==2.18.1
